+++ src/Enums/AlfColorRGBChannels.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Alf\Enums;
+
+enum AlfColorRGBChannels: string {
+
+    case R = 'r';
+    case G = 'g';
+    case B = 'b';
+
+}