+++ src/Interfaces/Integers/AlfIntSet.php
@@ -3,10 +3,9 @@
 namespace Alf\Interfaces\Integers;
 
 use Alf\Attributes\AlfAttrParameterIsInt;
-use Alf\Types\Scalars\AlfInt;
 
 interface AlfIntSet {
 
-    public function setFromInt(#[AlfAttrParameterIsInt] AlfInt|int|null $value) : static;
+    public function setFromInt(#[AlfAttrParameterIsInt] AlfIntGet|int|null $value) : static;
 
 }