+++ src/Interfaces/Integers/AlfIntWork.php
@@ -3,15 +3,14 @@
 namespace Alf\Interfaces\Integers;
 
 use Alf\Attributes\AlfAttrParameterIsInt;
-use Alf\Types\Scalars\AlfInt;
 
 interface AlfIntWork extends AlfIntGet, AlfIntSet {
 
-    public function add(#[AlfAttrParameterIsInt] AlfInt|int|null $value) : static;
+    public function add(#[AlfAttrParameterIsInt] AlfIntGet|int|null $value) : static;
 
     public function inc() : static;
 
-    public function sub(#[AlfAttrParameterIsInt] AlfInt|int|null $value) : static;
+    public function sub(#[AlfAttrParameterIsInt] AlfIntGet|int|null $value) : static;
 
     public function dec() : static;