+++ src/Interfaces/Strings/AlfStringGetTrait.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace Alf\Interfaces\Strings;
+
+use Alf\Attributes\AlfAttrAutoComplete;
+use Alf\Attributes\AlfAttrTraitAutoCall;
+use Alf\Services\AlfProgramming;
+use JetBrains\PhpStorm\Pure;
+
+trait AlfStringGetTrait {
+
+    /** @AlfAttrAutoComplete */
+    #[AlfAttrAutoComplete]
+    final public static function _AlfStringGet($obj) : AlfStringGet {
+        return AlfProgramming::_()->unused($obj);
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfStringGetTraitCTor() : void {
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfStringGetTraitDTor() : void {
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfStringGetTraitClone() : void {
+    }
+
+    #[Pure]
+    abstract public function getAsString() : string;
+
+    #[Pure]
+    public function __toString() : string {
+        return $this->getAsString();
+    }
+
+}