+++ src/Interfaces/Strings/AlfStringSetTrait.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Alf\Interfaces\Strings;
+
+use Alf\Attributes\AlfAttrAutoComplete;
+use Alf\Attributes\AlfAttrParameterIsString;
+use Alf\Attributes\AlfAttrTraitAutoCall;
+use Alf\Services\AlfProgramming;
+use Stringable;
+
+trait AlfStringSetTrait {
+
+    /** @AlfAttrAutoComplete */
+    #[AlfAttrAutoComplete]
+    final public static function _AlfStringSet($obj) : AlfStringSet {
+        return AlfProgramming::_()->unused($obj);
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfStringSetTraitCTor() : void {
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfStringSetTraitDTor() : void {
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfStringSetTraitClone() : void {
+    }
+
+    abstract public function setFromString(#[AlfAttrParameterIsString] AlfStringGet|Stringable|string|null $value) : static;
+
+}