+++ src/Interfaces/Strings/AlfStringWork.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Alf\Interfaces\Strings;
+
+interface AlfStringWork extends AlfStringGet, AlfStringSet {
+
+}