+++ src/Interfaces/Strings/AlfStringWorkTrait.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Alf\Interfaces\Strings;
+
+use Alf\Attributes\AlfAttrAutoComplete;
+use Alf\Attributes\AlfAttrTraitAutoCall;
+use Alf\Services\AlfProgramming;
+
+trait AlfStringWorkTrait {
+
+    use AlfStringGetTrait, AlfStringSetTrait;
+
+    /** @AlfAttrAutoComplete */
+    #[AlfAttrAutoComplete]
+    final public static function _AlfStringWork($obj) : AlfStringWork {
+        return AlfProgramming::_()->unused($obj, static::_AlfStringGet($obj), static::_AlfStringSet($obj));
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfStringWorkTraitCTor() : void {
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfStringWorkTraitDTor() : void {
+    }
+
+    /** @AlfAttrTraitAutoCall */
+    #[AlfAttrTraitAutoCall]
+    protected function _AlfStringWorkTraitClone() : void {
+    }
+
+}